Monitor state persistence fix (#43)
* Wrapped CanNotContinue errors for DKG

* Fixed logging of sync status

* Persist monitor state and load on startup.

* Corrected monitor db path

* Fixed spelling

* Removed ETHDKG phase length override. Fixed DB from earlier merge.

* Removed leftover commit/wait

+++ config/settings.go
@@ -39,7 +39,7 @@ type ethereumConfig struct {
 	RegistryAddress      string
 	RetryCount           int
 	RetryDelay           time.Duration
-	StartingBlock        int
+	StartingBlock        uint64
 	TestEther            string
 	Timeout              time.Duration
 }
